feat: add selectedItemSlice
