+++ src/components/result-component/resultsComponent.module.css
@@ -1,19 +1,26 @@
-.cardsContainer {
+.wrapper {
   display: flex;
   flex-direction: column;
   align-items: center;
   width: 40%;
 }
 
+.cardsContainer {
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  grid-auto-rows: 1fr;
+  align-items: center;
+  gap: 10px;
+}
+
 .card {
   display: flex;
-  flex-direction: column;
   margin-bottom: 10px;
   background-color: antiquewhite;
   padding: 10px;
   border-radius: 10px;
   border-color: transparent;
-  width: 50%;
+  width: 200px;
   height: fit-content;
   cursor: pointer;
 }
